# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Application API
  description: Dev Blog application
  contact:
    name: API Support
    email: diman5178@gmail.com
  version: v1
...
servers:
  - url: http://localhost:8080
    description: Local service
security:
  - bearerAuthScheme: [ ]
...
paths:
...
  /v1/users/{userId}/articles/{articleId}/comments/{commentId}:
    get:
      tags:
        - User comments
      operationId: find_3
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
                  - type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentResponse'
    put:
      tags:
        - User comments
      operationId: update_2
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
                  - type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentResponse'
    delete:
      tags:
        - User comments
      operationId: delete_3
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
                  - type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                type: object
        "204":
          description: No Content
    patch:
      tags:
        - User comments
      summary: Hide or show the comment
      operationId: enable_1
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: enabled
          in: query
          required: true
          schema:
            type: boolean
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
                  - type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: OK
...
  /v1/users/{userId}/articles/{articleId}/comments:
    get:
      tags:
        - User comments
      operationId: findAll_3
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
                  - type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCommentResponse'
    post:
      tags:
        - User comments
      operationId: create_2
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                  - type: object
                    properties:
                      com.project.devblog.handler.apierror.ApiError:
                        $ref: '#/components/schemas/ApiError'
                  - type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                properties:
                  com.project.devblog.handler.apierror.ApiError:
                    $ref: '#/components/schemas/ApiError'
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                type: object
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentResponse'
...
components:
  schemas:
    ApiError:
      required:
        - error
      type: object
      properties:
        status:
          type: string
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        debugMessage:
          type: string
        subErrors:
          type: array
          items:
            $ref: '#/components/schemas/ApiSubError'
        error:
          type: string
      discriminator:
        propertyName: error
    ApiSubError:
      type: object
...
    CommentRequest:
      required:
        - message
      type: object
      properties:
        message:
          type: string
        receiverId:
          type: string
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
        authorNickname:
          type: string
        receiverNickname:
          type: string
        creationDate:
          type: string
          format: date-time
        articleId:
          type: integer
          format: int32
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PageCommentResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponse'
        sort:
          $ref: '#/components/schemas/Sort'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          $ref: '#/components/schemas/PageableObject'
        number:
          type: integer
          format: int32
        empty:
          type: boolean
...
  securitySchemes:
    bearerAuthScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT